+++ src/ts/components/core/chip/ChipGroupContext.tsx
@@ -1,0 +1,9 @@
+import React, { createContext } from "react";
+
+interface ChipGroupContextProps {
+    chipOnClick?: (e: React.MouseEvent<HTMLInputElement>) => void;
+}
+
+const ChipGroupContext = createContext<ChipGroupContextProps | null>(null);
+
+export default ChipGroupContext;