+++ src/ts/components/core/radio/Radio.tsx
@@ -8,6 +8,7 @@ import { BoxProps } from "props/box";
 import { DashBaseProps, PersistenceProps } from "props/dash";
 import { StylesApiProps } from "props/styles";
 import React from "react";
+import RadioGroupContext from "./RadioGroupContext";
 
 interface Props
     extends BoxProps,
@@ -49,12 +50,15 @@ const Radio = (props: Props) => {
         ...others
     } = props;
 
+    const { radioOnClick } = React.useContext(RadioGroupContext) || {};
+
     return (
         <MantineRadio
             data-dash-is-loading={
                 (loading_state && loading_state.is_loading) || undefined
             }
             onChange={(ev) => setProps({ checked: ev.currentTarget.checked })}
+            onClick={radioOnClick}
             {...others}
         />
     );